refactoring eleventy
