+++ .eleventy.js
@@ -3,78 +3,33 @@ import UpgradeHelper from "@11ty/eleventy-upgrade-help";
 import markdownIt from 'markdown-it';
 import markdownItAttrs from 'markdown-it-attrs';
 import { eleventyImageTransformPlugin } from "@11ty/eleventy-img";
-import esbuild from 'esbuild';
 import fs from "fs";
-import path from "path";
-import * as sass from "sass";
+
+import { year, convertTime } from './src/_11ty/shortcodes.js';
+import { convertTime as convertTimeFilter } from './src/_11ty/filters.js';
+import { buildJs, compileSass } from './src/_11ty/build-assets.js';
 
 export default function (eleventyConfig) {
 
-    // Add a build event to trigger esbuild
+    // Add build events
     eleventyConfig.on('eleventy.before', async () => {
-        await esbuild.build({
-            entryPoints: ['src/assets/js/index.js'],
-            bundle: true,
-            outfile: 'public/assets/js/bundle.js',
-            sourcemap: true, // Optional: for easier debugging
-            minify: process.env.ELEVENTY_RUN_MODE === 'build', // Minify only in production
-        });
-    });
-
-    // Compile Sass
-    eleventyConfig.on("eleventy.before", () => {
-        // Ensure the output directory exists
-        if (!fs.existsSync("public/assets/css")) {
-            fs.mkdirSync("public/assets/css", { recursive: true });
-        }
-
-        try {
-            let result = sass.compile("src/assets/css/style.scss", {
-                style: "compressed",
-                loadPaths: ["src/assets/css/"]
-            });
-            fs.writeFileSync("public/assets/css/style.css", result.css);
-        } catch (err) {
-            console.error("Sass compilation error:", err);
-        }
+        await buildJs();
+        compileSass();
     });
 
     eleventyConfig.addGlobalData("env", process.env);
 
-    // Passthrough copy the entire assets directory to the output.
-    // This will copy CSS, images, and any other static files.
-    // JavaScript files are handled separately by the esbuild step.
+    // Passthrough copy
     eleventyConfig.addPassthroughCopy("src/assets");
 
+    // Shortcodes
+    eleventyConfig.addShortcode('year', year);
+    eleventyConfig.addShortcode('convertTime', convertTime);
 
-    // Adding Spotify access
-    /* eleventyConfig.addFilter("getTrack", async function(spotifyData, req) {
-        const recentTrack = await getRecentlyPlayedTrack(req);
-        return recentTrack;
-    }); */
+    // Filters
+    eleventyConfig.addFilter('convertTime', convertTimeFilter);
 
-    eleventyConfig.addShortcode('year', () => {
-        return `${new Date().getFullYear()}`;
-    });
-
-    eleventyConfig.addShortcode('convertTime', (UTCDate) => {
-        const options = {
-            year: 'numeric',
-            month: 'long',
-            day: 'numeric',
-        }
-        return `${new Date(UTCDate).toLocaleDateString(undefined, options)}`;
-    });
-
-    eleventyConfig.addFilter('convertTime', (UTCDate) => {
-        const options = {
-            year: 'numeric',
-            month: 'long',
-            day: 'numeric',
-        }
-        return `${new Date(UTCDate).toLocaleDateString(undefined, options)}`;
-    });
-
+    // Markdown Configuration
     const markdownItOptions = {
         html: true,
         breaks: false,
@@ -84,14 +39,12 @@ export default function (eleventyConfig) {
     const markdownLib = markdownIt(markdownItOptions).use(markdownItAttrs);
     eleventyConfig.setLibrary('md', markdownLib);
 
+    // Plugins
     eleventyConfig.addPlugin(UpgradeHelper);
-
-
     eleventyConfig.addPlugin(eleventyImageTransformPlugin, {
         extensions: "html",
         formats: ["webp"],
         widths: ["auto"],
-
         defaultAttributes: {
             loading: "lazy",
             sizes: "100vw",
@@ -99,18 +52,14 @@ export default function (eleventyConfig) {
         },
     });
 
-    // This is a development-only feature.
-    // It should not run in a production build environment.
+    // Development Settings
     if (process.env.ELEVENTY_RUN_MODE === "serve") {
         eleventyConfig.setBrowserSyncConfig({
             callbacks: {
                 ready: function (err, bs) {
-
                     bs.addMiddleware("*", (req, res) => {
                         const content_404 = fs.readFileSync('public/404.html');
-                        // Add 404 http status code in request header.
                         res.writeHead(404, { "Content-Type": "text/html; charset=UTF-8" });
-                        // Provides the 404 content without redirect.
                         res.write(content_404);
                         res.end();
                     });
@@ -119,12 +68,12 @@ export default function (eleventyConfig) {
         });
     }
 
-
     return {
         dir: {
             input: "src",
             output: "public",
-            includes: "_includes"
+            includes: "_includes",
+            layouts: "_layouts"
         },
     };
 };
