+++ src/_11ty/filters.js
@@ -1,0 +1,8 @@
+export const convertTime = (UTCDate) => {
+    const options = {
+        year: 'numeric',
+        month: 'long',
+        day: 'numeric',
+    }
+    return `${new Date(UTCDate).toLocaleDateString(undefined, options)}`;
+};
